requirements.txt
setuptools>=69.0.0
web3==6.11.3
python-dotenv>=1.0.0
eth-account==0.9.0
eth-typing==3.5.1
eth-utils>=2.0.0
requests>=2.31.0
aiohttp==3.9.1
pandas>=2.1.0
numpy==1.24.3
python-json-logger>=2.0.7
tenacity>=8.2.0
flashbots>=1.0.0
scikit-learn>=1.3.0  # For ML-based bid optimization
pytest==7.4.3
pytest-asyncio==0.21.1  # For async test support
python-dpdk>=22.11.0  # For network optimization
psutil>=5.9.0  # For CPU and memory management
pyroute2>=0.7.0  # For network interface management
